requirements.txt
datasets==3.5.0
mlflow==2.21.3
mlflow_skinny==2.21.3
nltk==3.9.1
openai==1.76.0
pandas==2.2.3
python-dotenv==1.1.0
torch==2.6.0
transformers>=4.35.0
accelerate>=0.24.0